backend/requirements.txt
aiofiles==22.1.0
aiosqlite==0.20.0
anyio==3.7.1
appnope==0.1.4
arcgis==2.3.1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.8.1
asttokens==2.4.1
attrs==24.2.0
babel==2.16.0
backports.tarfile==1.2.0
beautifulsoup4==4.12.3
bleach==6.1.0
cachetools==5.5.0
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
comm==0.2.2
contourpy==1.2.1
cryptography==43.0.0
cycler==0.12.1
dask==2024.8.1
debugpy==1.8.5
decorator==5.1.1
defusedxml==0.7.1
Django==5.1
django-cors-headers==4.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
executing==2.0.1
fastjsonschema==2.20.0
fonttools==4.53.1
fqdn==1.5.1
fsspec==2024.6.1
geographiclib==2.0
geomet==1.1.0
gssapi==1.8.3
idna==3.8
importlib_metadata==8.4.0
inflection==0.5.1
ipykernel==6.29.5
ipython==8.26.0
ipython-genutils==0.2.0
ipywidgets==7.8.3
isoduration==20.11.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.0.2
jedi==0.19.1
Jinja2==3.1.4
json5==0.9.25
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter-client==6.1.12
jupyter-events==0.10.0
jupyter-server==1.24.0
jupyter-ydoc==0.2.5
jupyter_core==5.7.2
jupyter_server_fileid==0.9.2
jupyter_server_ydoc==0.8.0
jupyterlab==3.6.8
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==1.1.9
keyring==25.3.0
kiwisolver==1.4.5
krb5==0.6.0
locket==1.0.0
lxml==5.3.0
MarkupSafe==2.1.5
matplotlib==3.9.2
matplotlib-inline==0.1.7
mistune==3.0.2
more-itertools==10.4.0
nbclassic==1.1.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
notebook==6.5.7
notebook_shim==0.2.4
numpy==1.26.4
oauthlib==3.2.2
packaging==24.1
pandas==2.1.4
pandocfilters==1.5.1
parso==0.8.4
partd==1.4.2
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.2.2
prometheus_client==0.20.0
prompt_toolkit==3.0.47
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
puremagic==1.27
pyarrow==17.0.0
pycparser==2.22
Pygments==2.18.0
PyJWT==2.9.0
pylerc==4.0
pyparsing==3.1.4
pyshp==2.3.1
pyspnego==0.11.1
python-dateutil==2.9.0.post0
python-json-logger==2.0.7
pytz==2024.1
PyYAML==6.0.2
pyzmq==26.2.0
referencing==0.35.1
requests==2.31.0
requests-gssapi==1.3.0
requests-kerberos==0.15.0
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.20.0
Send2Trash==1.8.3
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
sqlparse==0.5.1
stack-data==0.6.3
terminado==0.18.1
tinycss2==1.3.0
toolz==0.12.1
tornado==6.4.1
traitlets==5.14.3
truststore==0.9.2
types-python-dateutil==2.9.0.20240821
typing_extensions==4.12.2
tzdata==2024.1
ujson==5.10.0
uri-template==1.3.0
uritemplate==4.1.1
urllib3==2.2.2
wcwidth==0.2.13
webcolors==24.8.0
webencodings==0.5.1
websocket-client==1.8.0
widgetsnbextension==3.6.8
y-py==0.6.2
ypy-websocket==0.8.4
zipp==3.20.1
